refactor: common 패키지 구분
