+++ src/main/java/roomescape/reservation/ui/dto/AvailableReservationTimeWebResponse.java
@@ -2,7 +2,7 @@ package roomescape.reservation.ui.dto;
 
 import lombok.AccessLevel;
 import lombok.experimental.FieldNameConstants;
-import roomescape.common.validate.Validator;
+import roomescape.common.utils.Validator;
 
 import java.time.LocalTime;
