+++ src/main/java/roomescape/theme/domain/ThemeDescription.java
@@ -5,7 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.experimental.FieldNameConstants;
-import roomescape.common.validate.Validator;
+import roomescape.common.utils.Validator;
 
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
