+++ src/main/java/roomescape/theme/infrastructure/H2ThemeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.common.jdbc.JdbcUtils;
+import roomescape.common.utils.JdbcUtils;
 import roomescape.theme.application.converter.ThemeConverter;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeId;
