+++ src/main/java/roomescape/theme/ui/dto/CreateThemeWebRequest.java
@@ -2,7 +2,7 @@ package roomescape.theme.ui.dto;
 
 import lombok.AccessLevel;
 import lombok.experimental.FieldNameConstants;
-import roomescape.common.validate.Validator;
+import roomescape.common.utils.Validator;
 
 @FieldNameConstants(level = AccessLevel.PRIVATE)
 public record CreateThemeWebRequest(
