+++ src/main/java/roomescape/theme/ui/dto/ThemeResponse.java
@@ -1,7 +1,7 @@
 package roomescape.theme.ui.dto;
 
 import lombok.experimental.FieldNameConstants;
-import roomescape.common.validate.Validator;
+import roomescape.common.utils.Validator;
 
 @FieldNameConstants
 public record ThemeResponse(
