+++ src/test/java/roomescape/common/validate/ValidatorTest.java
@@ -2,6 +2,7 @@ package roomescape.common.validate;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.common.utils.Validator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
